+++ uptimerobot/data_source_uptimerobot_account_test.go
@@ -3,7 +3,7 @@ package uptimerobot
 import (
 	"testing"
 
-	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 )
 
 func TestUptimeRobotDataSourceAccount_basic(t *testing.T) {