+++ uptimerobot/provider.go
@@ -10,8 +10,8 @@ package uptimerobot
 */
 
 import (
-	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/hashicorp/terraform/terraform"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 	"github.com/louy/terraform-provider-uptimerobot/uptimerobot/api"
 )